Remove the filename nonsense
